Rename payloads.payloads to payload.payload
Sources and payloads will be renamed to payload and source.
