+++ configure.ac
@@ -178,11 +178,11 @@ AC_CONFIG_FILES([Makefile
                  pyanaconda/modules/localization/Makefile
                  pyanaconda/modules/users/Makefile
                  pyanaconda/modules/payloads/Makefile
-                 pyanaconda/modules/payloads/payloads/Makefile
-                 pyanaconda/modules/payloads/payloads/dnf/Makefile
-                 pyanaconda/modules/payloads/payloads/dnf/packages/Makefile
-                 pyanaconda/modules/payloads/payloads/live_os/Makefile
-                 pyanaconda/modules/payloads/payloads/live_image/Makefile
+                 pyanaconda/modules/payloads/payload/Makefile
+                 pyanaconda/modules/payloads/payload/dnf/Makefile
+                 pyanaconda/modules/payloads/payload/dnf/packages/Makefile
+                 pyanaconda/modules/payloads/payload/live_os/Makefile
+                 pyanaconda/modules/payloads/payload/live_image/Makefile
                  pyanaconda/modules/payloads/base/Makefile
                  pyanaconda/modules/payloads/sources/Makefile
                  pyanaconda/modules/payloads/sources/live_os/Makefile
