+++ pyanaconda/modules/payloads/payload/Makefile.am
@@ -14,10 +14,10 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-SUBDIRS = packages
+SUBDIRS = dnf live_os live_image
 
 pkgpyexecdir = $(pyexecdir)/py$(PACKAGE_NAME)
-dnf_moduledir = $(pkgpyexecdir)/modules/payloads/payloads/dnf
+dnf_moduledir = $(pkgpyexecdir)/modules/payloads/payload
 dnf_module_PYTHON = $(srcdir)/*.py
 
 MAINTAINERCLEANFILES = Makefile.in
