+++ pyanaconda/modules/payloads/payload/dnf/dnf_interface.py
@@ -20,7 +20,7 @@
 from dasbus.server.interface import dbus_interface
 
 from pyanaconda.modules.common.constants.interfaces import PAYLOAD_DNF
-from pyanaconda.modules.payloads.payloads.payload_base_interface import PayloadBaseInterface
+from pyanaconda.modules.payloads.payload.payload_base_interface import PayloadBaseInterface
 
 
 @dbus_interface(PAYLOAD_DNF.interface_name)
