+++ pyanaconda/modules/payloads/payload/dnf/packages/packages_interface.py
@@ -23,7 +23,7 @@ from dasbus.typing import *  # pylint: disable=wildcard-import
 
 from pyanaconda.modules.common.constants.objects import PAYLOAD_PACKAGES
 from pyanaconda.modules.common.base import KickstartModuleInterfaceTemplate
-from pyanaconda.modules.payloads.payloads.dnf.packages.constants import MultilibPolicy
+from pyanaconda.modules.payloads.payload.dnf.packages.constants import MultilibPolicy
 
 
 @dbus_interface(PAYLOAD_PACKAGES.interface_name)
