+++ pyanaconda/modules/payloads/payload/live_image/Makefile.am
@@ -14,10 +14,8 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-SUBDIRS = dnf live_os live_image
-
 pkgpyexecdir = $(pyexecdir)/py$(PACKAGE_NAME)
-dnf_moduledir = $(pkgpyexecdir)/modules/payloads/payloads
+dnf_moduledir = $(pkgpyexecdir)/modules/payloads/payload/live_image
 dnf_module_PYTHON = $(srcdir)/*.py
 
 MAINTAINERCLEANFILES = Makefile.in
