+++ pyanaconda/modules/payloads/payload/live_image/installation.py
