+++ pyanaconda/modules/payloads/payload/live_image/live_image_interface.py
@@ -23,7 +23,7 @@ from dasbus.server.property import emits_properties_changed
 
 from pyanaconda.modules.common.constants.objects import PAYLOAD_LIVE_IMAGE
 from pyanaconda.modules.common.containers import TaskContainer
-from pyanaconda.modules.payloads.payloads.payload_base_interface import PayloadBaseInterface
+from pyanaconda.modules.payloads.payload.payload_base_interface import PayloadBaseInterface
 
 
 @dbus_interface(PAYLOAD_LIVE_IMAGE.interface_name)
