+++ pyanaconda/modules/payloads/payload/live_image/utils.py
