+++ pyanaconda/modules/payloads/payload/live_os/Makefile.am
@@ -15,7 +15,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 pkgpyexecdir = $(pyexecdir)/py$(PACKAGE_NAME)
-dnf_moduledir = $(pkgpyexecdir)/modules/payloads/payloads/live_image
+dnf_moduledir = $(pkgpyexecdir)/modules/payloads/payload/live_os
 dnf_module_PYTHON = $(srcdir)/*.py
 
 MAINTAINERCLEANFILES = Makefile.in
