+++ pyanaconda/modules/payloads/payload/live_os/live_os.py
@@ -26,12 +26,12 @@ from pyanaconda.core.constants import INSTALL_TREE
 from pyanaconda.modules.common.constants.objects import PAYLOAD_LIVE_OS
 from pyanaconda.modules.common.errors.payload import SourceSetupError, IncompatibleSourceError
 from pyanaconda.modules.payloads.constants import SourceType
-from pyanaconda.modules.payloads.payloads.payload_base import PayloadBase
+from pyanaconda.modules.payloads.payload.payload_base import PayloadBase
 from pyanaconda.modules.payloads.base.initialization import PrepareSystemForInstallationTask, \
     CopyDriverDisksFilesTask, SetUpSourcesTask, TearDownSourcesTask, UpdateBLSConfigurationTask
 from pyanaconda.modules.payloads.base.installation import InstallFromImageTask
 from pyanaconda.modules.payloads.base.utils import get_dir_size, get_kernel_version_list
-from pyanaconda.modules.payloads.payloads.live_os.live_os_interface import LiveOSInterface
+from pyanaconda.modules.payloads.payload.live_os.live_os_interface import LiveOSInterface
 
 from pyanaconda.anaconda_loggers import get_module_logger
 log = get_module_logger(__name__)
