+++ pyanaconda/modules/payloads/payload/live_os/live_os_interface.py
@@ -21,7 +21,7 @@ from dasbus.server.interface import dbus_interface, dbus_signal
 from dasbus.typing import *  # pylint: disable=wildcard-import
 
 from pyanaconda.modules.common.constants.objects import PAYLOAD_LIVE_OS
-from pyanaconda.modules.payloads.payloads.payload_base_interface import PayloadBaseInterface
+from pyanaconda.modules.payloads.payload.payload_base_interface import PayloadBaseInterface
 
 
 @dbus_interface(PAYLOAD_LIVE_OS.interface_name)
