+++ pyanaconda/modules/payloads/payload/payload_base.py
