+++ pyanaconda/modules/payloads/payload/payload_base_interface.py
