+++ tests/nosetests/pyanaconda_tests/module_payload_dnf_test.py
@@ -26,10 +26,10 @@ from pyanaconda.modules.common.constants.objects import PAYLOAD_PACKAGES
 from pyanaconda.modules.common.errors import InvalidValueError
 from pyanaconda.modules.payloads.payload import PayloadService
 from pyanaconda.modules.payloads.payload_interface import PayloadInterface
-from pyanaconda.modules.payloads.payloads.dnf.packages.packages import PackagesModule
-from pyanaconda.modules.payloads.payloads.dnf.packages.packages_interface import \
+from pyanaconda.modules.payloads.payload.dnf.packages.packages import PackagesModule
+from pyanaconda.modules.payloads.payload.dnf.packages.packages_interface import \
     PackagesInterface
-from pyanaconda.modules.payloads.payloads.dnf.packages.constants import TIMEOUT_UNSET, \
+from pyanaconda.modules.payloads.payload.dnf.packages.constants import TIMEOUT_UNSET, \
     RETRIES_UNSET, LANGUAGES_DEFAULT, LANGUAGES_NONE
